+++ src/main/java/ru/yandex/practicum/filmorate/controllers/FilmController.java
@@ -97,5 +97,4 @@ public class FilmController {
         log.info("Get all genres");
         return filmService.getAllGenres();
     }
-
 }
