+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -1,18 +1,13 @@
 package ru.yandex.practicum.filmorate.model;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
-import ru.yandex.practicum.filmorate.exceptions.ModelAlreadyExistException;
-import ru.yandex.practicum.filmorate.exceptions.ModelNotFoundException;
 import ru.yandex.practicum.filmorate.exceptions.ValidationException;
 
 import javax.validation.constraints.*;
 import java.time.LocalDate;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Map;
-import java.util.Set;
 
 @EqualsAndHashCode
 @Data
@@ -35,12 +30,6 @@ public class User {
     @NotNull
     private LocalDate birthday;
 
-    @JsonIgnore
-    private Set<Integer> friendsId;
-
-    @JsonIgnore
-    private Set<Integer> likedFilmsId;
-
     public User(String email, String login, String name, LocalDate birthday) {
         this.email = email;
         this.login = login;
@@ -50,37 +39,7 @@ public class User {
             this.name = name;
         }
         this.birthday = birthday;
-        this.friendsId = new HashSet<>();
-        this.likedFilmsId = new HashSet<>();
         validate();
-    }
-
-    public void addUserToFriend(int friendId) {
-        if (friendsId.contains(friendId)) {
-            throw new ModelAlreadyExistException("Users are already friends");
-        }
-        friendsId.add(friendId);
-    }
-
-    public void deleteUserFromFriend(int friendId) {
-        if (!friendsId.contains(friendId)) {
-            throw new ModelNotFoundException("Nothing to delete");
-        }
-        friendsId.remove(friendId);
-    }
-
-    public void addLike(int filmId) {
-        if (likedFilmsId.contains(filmId)) {
-            throw new ModelAlreadyExistException("Film already liked by user");
-        }
-        likedFilmsId.add(filmId);
-    }
-
-    public void deleteLike(int filmId) {
-        if (!likedFilmsId.contains(filmId)) {
-            throw new ModelNotFoundException("Nothing to delete");
-        }
-        likedFilmsId.remove(filmId);
     }
 
     public Map<String, Object> toMap() {
