+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -14,15 +14,7 @@ public interface FilmStorage {
 
     Collection<Film> getFilms();
 
-    Collection<MPA> getAllMpa();
-
-    Collection<Genre> getAllGenres();
-
     Film getFilmById(int id);
-
-    MPA getMpaById(int id);
-
-    Genre getGenreById(int id);
 
     void deleteFilmById(int id);
